# Conflicts:
#	src/app/layout/default-layout/boxes/search/cris-layout-search-box.component.html
#	src/app/search-page/search.component.html

+++ src/app/register-page/create-profile/create-profile.component.ts
@@ -127,7 +127,7 @@ export class CreateProfileComponent implements OnInit {
    */
   submitEperson() {
     if (!(this.userInfoForm.invalid || this.isInValidPassword)) {
-      const values = {
+      let values = {
         metadata: {
           'eperson.firstname': [
             {
@@ -156,15 +156,19 @@ export class CreateProfileComponent implements OnInit {
         requireCertificate: false
       };
 
-      // If the End User Agreement is accepted, add end-user agreement metadata to the user
-      if (this.userAgreementAccept) {
-        values.metadata[END_USER_AGREEMENT_METADATA_FIELD] = [
-          {
-            value: String(true)
-          }
-        ];
-        this.endUserAgreementService.removeCookieAccepted();
-      }
+      // If the End User Agreement cookie is accepted, add end-user agreement metadata to the user
+      this.endUserAgreementService.isUserAgreementEnabled().subscribe((isUserAgreementEnabled) => {
+        if (isUserAgreementEnabled && this.userAgreementAccept) {
+          values = Object.assign({}, values, Object.assign({}, values.metadata, {
+            [END_USER_AGREEMENT_METADATA_FIELD]: [
+              {
+                value: String(true)
+              }
+            ]
+          }));
+          this.endUserAgreementService.removeCookieAccepted();
+        }
+      });
 
       const eperson = Object.assign(new EPerson(), values);
       this.ePersonDataService.createEPersonForToken(eperson, this.token).pipe(
